requirements.txt
PyQt5
psycopg2
python-dotenv